+++ plugins/wpgraphql-logging/src/Logger/LoggerService.php
@@ -12,7 +12,7 @@ use Monolog\Processor\ProcessIdProcessor;
 use Monolog\Processor\ProcessorInterface;
 use Monolog\Processor\WebProcessor;
 use WPGraphQL\Logging\Logger\Handlers\WordPressDatabaseHandler;
-use WPGraphQL\Logging\Logger\Processors\WPGraphQLQueryProcessor;
+use WPGraphQL\Logging\Logger\Processors\RequestHeadersProcessor;
 
 /**
  * LoggerService class for managing the Monolog logger instance.
@@ -44,7 +44,7 @@ class LoggerService {
 	/**
 	 * The instance of the logger based off the channel name.
 	 *
-	 * @var array<LoggerService>
+	 * @var array<\WPGraphQL\Logging\Logger\LoggerService>
 	 */
 	protected static array $instances = [];
 
@@ -90,16 +90,16 @@ class LoggerService {
 		?array $processors = null,
 		?array $default_context = null
 	): LoggerService {
-		if ( isset(self::$instances[$channel]) ) {
-			return self::$instances[$channel];
+		if ( isset( self::$instances[ $channel ] ) ) {
+			return self::$instances[ $channel ];
 		}
 
 		$processors      = $processors ?? self::get_default_processors();
 		$handlers        = $handlers ?? self::get_default_handlers();
 		$default_context = $default_context ?? self::get_default_context();
 
-		self::$instances[$channel] = new self( $channel, $handlers, $processors, $default_context );
-		return self::$instances[$channel];
+		self::$instances[ $channel ] = new self( $channel, $handlers, $processors, $default_context );
+		return self::$instances[ $channel ];
 	}
 
 	/**
@@ -224,8 +224,10 @@ class LoggerService {
 			new MemoryPeakUsageProcessor(), // Logs memory peak data.
 			new WebProcessor(), // Logs web request data. e.g. IP address, request method, URI, etc.
 			new ProcessIdProcessor(), // Logs the process ID.
-			new WPGraphQLQueryProcessor(), // Custom processor to capture GraphQL request data.
+			new RequestHeadersProcessor(), // Custom processor to capture request headers.
 		];
+
+		// Filter for users to add their own processors.
 		return apply_filters( 'wpgraphql_logging_default_processors', $default_processors );
 	}
 
@@ -240,6 +242,8 @@ class LoggerService {
 		$default_handlers = [
 			new WordPressDatabaseHandler(),
 		];
+
+		// Filter for users to add their own handlers.
 		return apply_filters( 'wpgraphql_logging_default_handlers', $default_handlers );
 	}
 
@@ -256,6 +260,7 @@ class LoggerService {
 			'site_url'       => home_url(),
 		];
 
+		// Filter for users to modify the default context.
 		return apply_filters( 'wpgraphql_logging_default_context', $context );
 	}
 }